fix create table

+++ create_tables.sql
@@ -2,7 +2,7 @@ CREATE TABLE item
 (
 	itemID INTEGER PRIMARY KEY,
 	name CHAR(20),
-	price DOUBLE(4, 2),
+	price NUMBER(6, 2),
 	type CHAR(20)
 );
